move compat code to reth_compat
